adjust how backtracks are counted
Include cells that were solved (and thus didn't have an associated
transformation) when backtracking. Adjustment made so that both ss
and ss-opt count backtracks the same way. In ss-opt, the cells in the
worklist are those that weren't part of the initial puzzle. ss does
not distinguish using a cell's initial state, and formerly skipped
all completed cells regardless of reason. Since the worklist for ss
contains all cells in the puzzle, this change will reflect that.